Moved menu functions to separate file MenuFunctions.c so the menu code is more focused on just handling the menu.
